packages/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "ALPS Writer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "awscli>=1.37.12",
    "boto3>=1.36.12",
    "chainlit==2.6.6",
    "markdown>=3.7",
    "pdfplumber>=0.11.5",
    "python-dotenv>=1.0.1",
    "strands-agents>=1.3.0",
    "structlog>=25.2.0",
    "tavily-python>=0.5.0",
    "tiktoken>=0.9.0",
]

[tool.ruff]
ignore = ["E402"]

[tool.ruff.isort]
known-first-party = ["dotenv"]
known-local-folder = ["src"]
section-order = ["future", "standard-library", "first-party", "third-party", "local-folder"]
